Do not repeat has_cufile
